+++ src/ts/components/tooltip.ts
@@ -164,7 +164,7 @@ export class Tooltip extends Component<TooltipOptions> {
     tooltipContentEl.classList.add('tooltip-content');
     this._setTooltipContent(tooltipContentEl);
     this.tooltipEl.appendChild(tooltipContentEl);
-    document.body.appendChild(this.tooltipEl);
+    Utils.portalRoot(this.el).appendChild(this.tooltipEl);
   }
 
   _setTooltipContent(tooltipContentEl: HTMLElement) {